requirements.txt
pandas==2.2.1
requests==2.31.0
python-dotenv==1.0.1
jsonschema==4.22.0
python-jose==3.3.0
passlib==1.7.4
SQLAlchemy==2.0.27
pydantic==2.7.4
fastapi==0.111.0
pytest==7.4.0
uvicorn==0.30.1
bcrypt==4.0.1